updated to fix clang compiling issues

+++ src/emannlib-common/Singleton.hpp
@@ -17,8 +17,8 @@ namespace emannlib
 		Singleton()
 		{
 			GD_ASSERT(ms_Singleton == NULL);
-			int offset = (int) (T*) 1 - (int) (Singleton <T>*)(T*) 1;
-			ms_Singleton = (T*) ((int)this + offset);
+			int offset = (long int) (T*) 1 - (long int) (Singleton <T>*)(T*) 1;
+			ms_Singleton = (T*) ((long int)this + offset);
 		}
 		~Singleton()
 		{
